+++ ios/Adjust/Adjust.m
@@ -30,10 +30,13 @@ NSString * const ADJAdRevenueSourceAdMost = @"admost_sdk";
 NSString * const ADJAdRevenueSourceUnity = @"unity_sdk";
 NSString * const ADJAdRevenueSourceHeliumChartboost = @"helium_chartboost_sdk";
 NSString * const ADJAdRevenueSourcePublisher = @"publisher_sdk";
+NSString * const ADJAdRevenueSourceTopOn = @"topon_sdk";
+NSString * const ADJAdRevenueSourceADX = @"adx_sdk";
 
 NSString * const ADJUrlStrategyIndia = @"UrlStrategyIndia";
 NSString * const ADJUrlStrategyChina = @"UrlStrategyChina";
 NSString * const ADJUrlStrategyCn = @"UrlStrategyCn";
+NSString * const ADJUrlStrategyCnOnly = @"UrlStrategyCnOnly";
 
 NSString * const ADJDataResidencyEU = @"DataResidencyEU";
 NSString * const ADJDataResidencyTR = @"DataResidencyTR";
@@ -49,6 +52,8 @@ NSString * const ADJDataResidencyUS = @"DataResidencyUS";
 @property (nonatomic, strong) id<ADJActivityHandler> activityHandler;
 
 @property (nonatomic, strong) ADJSavedPreLaunch *savedPreLaunch;
+
+@property (nonatomic) AdjustResolvedDeeplinkBlock cachedResolvedDeeplinkBlock;
 
 @end
 
@@ -75,6 +80,7 @@ static dispatch_once_t onceToken = 0;
     self.activityHandler = nil;
     self.logger = [ADJAdjustFactory logger];
     self.savedPreLaunch = [[ADJSavedPreLaunch alloc] init];
+    self.cachedResolvedDeeplinkBlock = nil;
     return self;
 }
 
@@ -123,6 +129,13 @@ static dispatch_once_t onceToken = 0;
     }
 }
 
++ (void)processDeeplink:(nonnull NSURL *)deeplink
+      completionHandler:(void (^_Nonnull)(NSString * _Nonnull resolvedLink))completionHandler {
+    @synchronized (self) {
+        [[Adjust getInstance] processDeeplink:deeplink completionHandler:completionHandler];
+    }
+}
+
 + (void)setDeviceToken:(NSData *)deviceToken {
     @synchronized (self) {
         [[Adjust getInstance] setDeviceToken:[deviceToken copy]];
@@ -148,6 +161,12 @@ static dispatch_once_t onceToken = 0;
 + (NSString *)idfa {
     @synchronized (self) {
         return [[Adjust getInstance] idfa];
+    }
+}
+
++ (NSString *)idfv {
+    @synchronized (self) {
+        return [[Adjust getInstance] idfv];
     }
 }
 
@@ -347,9 +366,9 @@ static dispatch_once_t onceToken = 0;
         [self.logger error:@"Adjust already initialized"];
         return;
     }
-    self.activityHandler = [[ADJActivityHandler alloc]
-                                initWithConfig:adjustConfig
-                                savedPreLaunch:self.savedPreLaunch];
+    self.activityHandler = [[ADJActivityHandler alloc] initWithConfig:adjustConfig
+                                                       savedPreLaunch:self.savedPreLaunch
+                                           deeplinkResolutionCallback:self.cachedResolvedDeeplinkBlock];
 }
 
 - (void)trackEvent:(ADJEvent *)event {
@@ -400,6 +419,27 @@ static dispatch_once_t onceToken = 0;
     [self.activityHandler appWillOpenUrl:url withClickTime:clickTime];
 }
 
+- (void)processDeeplink:(nonnull NSURL *)deeplink
+      completionHandler:(void (^_Nonnull)(NSString * _Nonnull resolvedLink))completionHandler {
+    // if resolution result is not wanted, fallback to default method
+    if (completionHandler == nil) {
+        [self appWillOpenUrl:deeplink];
+        return;
+    }
+    // if deep link processing is triggered prior to SDK being initialized
+    [ADJUserDefaults cacheDeeplinkUrl:deeplink];
+    NSDate *clickTime = [NSDate date];
+    if (![self checkActivityHandler]) {
+        [ADJUserDefaults saveDeeplinkUrl:deeplink andClickTime:clickTime];
+        self.cachedResolvedDeeplinkBlock = completionHandler;
+        return;
+    }
+    // if deep link processing was triggered with SDK being initialized
+    [self.activityHandler processDeeplink:deeplink
+                                clickTime:clickTime
+                        completionHandler:completionHandler];
+}
+
 - (void)setDeviceToken:(NSData *)deviceToken {
     [ADJUserDefaults savePushTokenData:deviceToken];
 
@@ -432,6 +472,10 @@ static dispatch_once_t onceToken = 0;
 
 - (NSString *)idfa {
     return [ADJUtil idfa];
+}
+
+- (NSString *)idfv {
+    return [ADJUtil idfv];
 }
 
 - (NSURL *)convertUniversalLink:(NSURL *)url scheme:(NSString *)scheme {